authority/requirements.txt
djangorestframework
markdown   
django-filter  
django
PyJWT
jsonschema
